LayoutTests/platform/wpe/webrtc/sframe-test-vectors-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

FAIL Decryption test 1 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 1 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 2 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 2 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 3 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 3 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 4 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 4 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 5 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 5 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 6 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 6 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 7 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 7 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 8 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 8 for ciphersuite 1 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 1 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 1 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 2 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 2 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 3 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 3 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 4 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 4 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 5 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 5 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 6 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 6 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 7 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 7 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Decryption test 8 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
FAIL Encryption/Decryption test 8 for ciphersuite 2 assert_equals: expected "46726f6d2068656176656e6c79206861726d6f6e79202f2f205468697320756e6976657273616c206672616d6520626567616e" but got ""
